Revert "Re-implement `Tooltip` component (#1291)"
This reverts commit 5db294995edb2b1d5d29637502ff241dfb3996a5.
