+++ packages/bezier-react/package.json
@@ -140,7 +140,6 @@
     "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-toolbar": "^1.0.4",
-    "@radix-ui/react-tooltip": "^1.0.6",
     "@radix-ui/react-visually-hidden": "^1.0.3",
     "classnames": "^2.3.2",
     "react-resize-detector": "^9.1.0",
