+++ packages/bezier-react/src/components/KeyValueListItem/common/ItemActions/ItemAction.styled.ts
@@ -8,7 +8,7 @@ import { isNil } from '~/src/utils/typeUtils'
 
 import { Icon } from '~/src/components/Icon'
 import { LegacyIcon } from '~/src/components/LegacyIcon'
-import { LegacyTooltip } from '~/src/components/LegacyTooltip'
+import { Tooltip } from '~/src/components/Tooltip'
 
 export const ItemActionWrapper = styled.div`
   display: flex;
@@ -41,7 +41,7 @@ export const ActionIconWrapper = styled.div<ActionWrapperProps>`
         color: ${foundation?.theme?.[hoverIconColor]};
       `}
     }
-
+    
     ${ActionLegacyIcon} {
       ${({ foundation, hoverIconColor }) => !isNil(hoverIconColor) && css`
         color: ${foundation?.theme?.[hoverIconColor]};
@@ -50,6 +50,6 @@ export const ActionIconWrapper = styled.div<ActionWrapperProps>`
   }
 `
 
-export const ActionIconTooltip = styled(LegacyTooltip)`
+export const ActionIconTooltip = styled(Tooltip)`
   margin-left: auto;
 `
