--- packages/bezier-react/src/components/LegacyTooltip/index.ts
@@ -1,5 +1,0 @@
-export { LegacyTooltip } from './LegacyTooltip'
-
-export { LegacyTooltipPosition } from './LegacyTooltip.types'
-
-export type { LegacyTooltipProps } from './LegacyTooltip.types'
