+++ packages/bezier-react/src/components/Tooltip/index.ts
@@ -1,12 +1,12 @@
-export {
-  Tooltip,
-  TooltipProvider,
-} from './Tooltip'
-
-export { TooltipPosition } from './Tooltip.types'
+import Tooltip from './Tooltip'
+import { TooltipPosition } from './Tooltip.types'
+import type TooltipProps from './Tooltip.types'
 
 export type {
   TooltipProps,
-  TooltipProviderProps,
-} from './Tooltip.types'
+}
 
+export {
+  Tooltip,
+  TooltipPosition,
+}
