+++ packages/bezier-react/src/index.ts
@@ -56,7 +56,6 @@ export * from '~/src/components/AlphaSmoothCornersBox'
 
 export * from '~/src/components/LegacyIcon'
 export * from '~/src/components/LegacyRadio'
-export * from '~/src/components/LegacyTooltip'
 export * from '~/src/components/LegacySegmentedControl'
 
 /* Hooks for Component */
