+++ packages/bezier-react/src/providers/BezierProvider.tsx
@@ -9,8 +9,6 @@ import {
   type ThemeVarsAdditionalType,
 } from '~/src/foundation'
 
-import { TooltipProvider } from '~/src/components/Tooltip'
-
 interface BezierProviderProps {
   foundation: Foundation & GlobalStyleProp
   children: React.ReactNode
@@ -24,14 +22,12 @@ function BezierProvider({
 }: BezierProviderProps) {
   return (
     <FoundationProvider foundation={foundation}>
-      <TooltipProvider>
-        <GlobalStyle foundation={foundation} />
-        <ThemeVars
-          foundation={foundation}
-          scope={themeVarsScope}
-        />
-        { children }
-      </TooltipProvider>
+      <GlobalStyle foundation={foundation} />
+      <ThemeVars
+        foundation={foundation}
+        scope={themeVarsScope}
+      />
+      { children }
     </FoundationProvider>
   )
 }
